+++ routes/web.php
@@ -122,6 +122,7 @@ Route::prefix('/carte')->name('map.')->group(function () {
 });
 
 Route::get('/equipe-et-partenaires', 'PublicController@teamPage')->name('team');
+Route::get('/à-propos', 'PublicController@aboutPage')->name('about');
 
 Route::get('/mrc/json', 'RcmController@listJson')->name('rcm.list-json');
 Route::get('/mrc/json/{region}', 'RcmController@listJson')->name('rcm.list-json-region');