+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Next
 - Changed icon for close and show place.
+- Moved about text to its own page.
 
 # 2.8.0
 - Fixed page title in social share meta.